requirements.txt
pillow==9.5.0
black==23.3.0
numpy==1.24.2
